backend/requirements.txt
fastapi==0.108.0
uvicorn==0.27.0
python-dotenv==1.0.0
pymilvus==2.3.3
langchain==0.1.5
langchain-community==0.0.29
langchain-core==0.1.33
pdfplumber==0.10.4
pytesseract==0.3.10
pillow==10.3.0
beautifulsoup4==4.12.3
requests==2.31.0
validators==0.22.0
sentence-transformers==2.2.2
deepseek-ai>=0.1.0